requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiohttp==3.9.1
openai==1.3.5
pydantic==2.5.2
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
selenium==4.15.2
scikit-learn==1.3.2
nltk==3.8.1
lightgbm==4.1.0
numpy==1.24.3
pandas==2.0.3
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
tqdm==4.66.1
scikit-learn==1.3.2
nltk==3.8.1 
